lab 12/instagram-ai-lab12/ai-filter/requirements.txt
flask
pillow
scikit-image
requests
flask-cors
